forecast-asset/forecast-asset-movement/requirements.txt
functions-framework==3.*
db-dtypes
pandas
numpy
scikit-learn
joblib
google-cloud-bigquery
google-cloud-storage
tensorflow==2.11.0

